"Enhance grid layout and adjust modal actions
Updated templates/gridPage.html to create a more dynamic and visually engaging grid layout by adjusting the rows, columns, and positioning of grid components. This change will improve usability and aesthetics of the site.

Adjusted templates/modals.html and templates/sidebar.html to handle login events differently. Instead of the previous closeModal and click combination, a new reloadMain action was added to refresh the main content after a user logs in. This was necessary to ensure all user-specific content is properly loaded after sign in."